[고보경] 웹소켓 설정진행중입니다

+++ dbdevdeep/src/main/java/com/dbdevdeep/chat/vo/ChatMsgVo.java
@@ -16,8 +16,8 @@ import lombok.Setter;
 @ToString
 @Builder
 public class ChatMsgVo {
-	private Long msg_no;
-	private Long room_no;
+	private int msg_no;
+	private int room_no;
 	private String writer_id;
 	private String chat_content;
 	private int is_important;
